# apps/worker/docs/swagger.yaml
...
info:
  contact: {}
  description: LLM API Gateway — aggregate providers, manage load balancing, and track
    usage/costs.
  title: Wheel API
  version: 1.0.0
paths:
  /api/v1/apikey/create:
    post:
      consumes:
        - application/json
      parameters:
        - description: "API key config: name, expireAt, maxCost, supportedModels"
          in: body
          name: body
          required: true
          schema:
            type: object
      produces:
        - application/json
      responses:
        "200":
          description: "{success: true, data: APIKey}"
          schema:
            type: object
        "400":
          description: "{success: false, error: string}"
          schema:
            type: object
      security:
        - BearerAuth: []
      summary: Create a new API key
      tags:
        - API Keys
  /api/v1/apikey/delete/{id}:
    delete:
      parameters:
        - description: API Key ID
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: "{success: true}"
          schema:
            type: object
        "400":
          description: "{success: false, error: string}"
          schema:
            type: object
      security:
        - BearerAuth: []
      summary: Delete an API key
      tags:
        - API Keys
  /api/v1/apikey/list:
    get:
      produces:
        - application/json
      responses:
        "200":
          description: "{success: true, data: {apiKeys: []APIKey}}"
          schema:
            type: object
      security:
        - BearerAuth: []
      summary: List all API keys
      tags:
        - API Keys
  /api/v1/apikey/update:
    post:
      consumes:
        - application/json
      parameters:
        - description: Partial API key fields to update (id required)
          in: body
          name: body
          required: true
          schema:
            type: object
      produces:
        - application/json
      responses:
        "200":
          description: "{success: true}"
          schema:
            type: object
        "400":
          description: "{success: false, error: string}"
          schema:
            type: object
      security:
        - BearerAuth: []
      summary: Update an API key
      tags:
        - API Keys
...
  /api/v1/user/apikey/login:
    get:
      produces:
        - application/json
      responses:
        "200":
          description: "{success: true, data: {id, name, enabled, expireAt, maxCost,
            totalCost, supportedModels}}"
          schema:
            type: object
        "404":
          description: "{success: false, error: string}"
          schema:
            type: object
      security:
        - ApiKeyAuth: []
      summary: Get API key info (end-user)
      tags:
        - API Keys
  /api/v1/user/apikey/stats:
    get:
      produces:
        - application/json
      responses:
        "200":
          description: "{success: true, data: {id, name, totalCost, maxCost, enabled,
            expireAt}}"
          schema:
            type: object
        "404":
          description: "{success: false, error: string}"
          schema:
            type: object
      security:
        - ApiKeyAuth: []
      summary: Get API key usage stats (end-user)
      tags:
        - API Keys
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"